Fixes #378: option added to edit the sub project (#388)
