+++ views/dashboard.jade
@@ -1,7 +1,9 @@
 extends layout
 
 block content
+  script(src="/scripts/search.js")
   script(src="/scripts/socket-dashboard.js")
+  script(src="/scripts/notification.js")
   nav.navbar.navbar-default
     .container
       .navbar-header
