+++ views/repository/layout.jade
@@ -11,6 +11,8 @@ html
     script(src='https://code.jquery.com/jquery-1.11.1.js')
     script(src='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js')
   body
+    #notification-container
+      li.notification Placeholder
     include ../template/header
     block content
     include ../template/footer
